Refactor operation class
- Refactor all-books to add and list books
- Refactor all-peoples to add and list people
- Refactor all-rentals to add and list rentals
