+++ all_rentals.rb
@@ -1,58 +1,15 @@
-require './app'
 require './check_input'
 
-class Operation
+class AllRentals
   include CheckInput
 
-  def initialize
-    @app = App.new
+  def initialize(app)
+    @app = app
     @option = ''
-    @type = ''
-    @age = ''
-    @name = ''
-    @permission = ''
-    @specialization = ''
-    @title = ''
-    @author = ''
     @book = ''
     @person = ''
     @date = ''
     @id = ''
-  end
-
-  def list_all_books
-    puts 'The book list is empty!' if @app.list_books.length.zero?
-    @app.list_books.each { |book| puts "Title: \"#{book.title}\", Author: #{book.author}" }
-  end
-
-  def list_all_people
-    puts 'The people list is empty!' if @app.list_people.length.zero?
-    @app.list_people.each do |people|
-      puts "[#{people.class}] Name: #{people.name}, ID: #{people.id}, Age: #{people.age}"
-    end
-  end
-
-  def create_a_person
-    @type = check_input('Do you want to create a student (1) or a teacher (2)? [Input the number]: ') do
-      %w[1 2].include?(@option)
-    end
-    @age = check_input('Age: ') { @option.match?(/^\d+$/) }
-    @name = check_input('Name: ') { @option != '' }
-    if @type == '1'
-      @permission = check_input('Has parent permission? [Y/N]: ') { %w[y n].include?(@option.downcase) }
-      @app.create_new_student(@age, @name, @permission)
-    else
-      @specialization = check_input('Specialization: ') { @option != '' }
-      @app.create_new_teacher(@specialization, @age, @name)
-    end
-    puts 'Person created successfully'
-  end
-
-  def create_a_book
-    @title = check_input('Title: ') { @option != '' }
-    @author = check_input('Author: ') { @option != '' }
-    @app.create_new_book(@title, @author)
-    puts 'Book created successfully'
   end
 
   def select_book
